separate graphml processing functions
