+++ src/main/data-managers/ProtocolManager.js
@@ -8,8 +8,8 @@ const ProtocolDB = require('./ProtocolDB');
 const SessionDB = require('./SessionDB');
 const { ErrorMessages, RequestError } = require('../errors/RequestError');
 const { readFile, rename, tryUnlink } = require('../utils/promised-fs');
+const { validateGraphML, convertGraphML } = require('../utils/importGraphML');
 const { hexDigest } = require('../utils/sha256');
-const dom = require('xmldom');
 const { sendToGui } = require('../guiProxy');
 
 const validProtocolFileExts = ['netcanvas'];
@@ -35,7 +35,6 @@ class ProtocolManager {
     this.presentImportProtocolDialog = this.presentImportProtocolDialog.bind(this);
     this.presentImportSessionDialog = this.presentImportSessionDialog.bind(this);
     this.validateAndImport = this.validateAndImport.bind(this);
-    this.domparser = new dom.DOMParser();
 
     const dbFile = path.join(dataDir, 'db', 'protocols.db');
     this.db = new ProtocolDB(dbFile);
@@ -392,7 +391,7 @@ class ProtocolManager {
       sendToGui(emittedEvents.SESSIONS_IMPORT_STARTED);
 
       return (this.fileContents(userFilepath, ''))
-        .then(bufferContents => this.validateGraphML(bufferContents))
+        .then(bufferContents => validateGraphML(bufferContents))
         .then(xmlDoc => this.processGraphML(xmlDoc))
         .then(({ protocolId, sessions }) => sessions.map(
           session => this.addSessionData(protocolId, session)))
@@ -448,35 +447,6 @@ class ProtocolManager {
       });
   }
 
-  validateGraphML(bufferContents) {
-    const xmlDoc = this.domparser.parseFromString(bufferContents.toString(), 'text/xml');
-
-    // basic header validation
-    const graphml = xmlDoc.getElementsByTagName('graphml');
-    if (!graphml || !graphml[0] || graphml[0].getAttribute('xmlns:nc') !== 'http://schema.networkcanvas.com/xmlns' ||
-      graphml[0].getAttribute('xmlns') !== 'http://graphml.graphdrawing.org/xmlns'
-    ) {
-      throw new RequestError(`${ErrorMessages.InvalidSessionFormat}: missing headers`);
-    }
-
-    // all graphs within the graphml must have the same protocol
-    const graphs = graphml[0].getElementsByTagName('graph');
-    if (graphs < 1) {
-      throw new RequestError(`${ErrorMessages.InvalidSessionFormat}: missing graph`);
-    }
-    let graphmlProtocolId;
-    Array.from(graphs).forEach((graph) => {
-      const protocolId = graph.getAttribute('nc:remoteProtocolID');
-      if (!graphmlProtocolId) {
-        graphmlProtocolId = protocolId;
-      } else if (graphmlProtocolId !== protocolId) {
-        throw new RequestError(`${ErrorMessages.InvalidSessionFormat}: may only contain one protocol`);
-      }
-    });
-
-    return xmlDoc;
-  }
-
   processGraphML(xmlDoc) {
     const graphml = xmlDoc.getElementsByTagName('graphml');
     const graphs = graphml[0].getElementsByTagName('graph');
@@ -488,161 +458,8 @@ class ProtocolManager {
         if (!protocol) {
           throw new RequestError(`${ErrorMessages.ProtocolNotFoundForSession}: ${protocolName}`);
         }
-        return this.convertGraphML(xmlDoc, protocol);
+        return convertGraphML(xmlDoc, protocol);
       });
-  }
-
-  convertGraphML(xmlDoc, protocol) {
-    const graphml = xmlDoc.getElementsByTagName('graphml');
-
-    const sessions = [];
-    const graphs = graphml[0].getElementsByTagName('graph');
-    const protocolId = graphs && graphs[0].getAttribute('nc:remoteProtocolID');
-    Array.from(graphs).forEach((graph) => {
-      // process session variables
-      const session = {};
-      const sessionId = graph.getAttribute('nc:sessionUUID');
-      session.uuid = sessionId;
-      session.data = { sessionVariables: {
-        sessionId,
-        caseId: graph.getAttribute('nc:caseId'),
-        remoteProtocolID: protocolId,
-        protocolName: graph.getAttribute('nc:protocolName'),
-        sessionExported: graph.getAttribute('nc:sessionExportTime'),
-        sessionStart: graph.getAttribute('nc:sessionStartTime'),
-        sessionFinish: graph.getAttribute('nc:sessionFinishTime'),
-      } };
-
-      const entityElements = graph.childNodes;
-      session.data.ego = {};
-      session.data.ego.attributes = {};
-      session.data.nodes = [];
-      session.data.edges = [];
-      for (let i = 0; i < entityElements.length; i += 1) {
-        if (entityElements[i].nodeType === 1) {
-          switch (entityElements[i].tagName) {
-            case 'data': // process ego
-              session.data.ego = this.processVariable(
-                entityElements[i], session.data.ego, xmlDoc, protocol.codebook.ego);
-              break;
-            // eslint-disable-next-line no-case-declarations
-            case 'node':
-              const nodeElement = entityElements[i].childNodes;
-              let node = {};
-              node.attributes = {};
-              const nodeType = this.lookUpEntityType(nodeElement, protocol.codebook.node);
-              Array.from(nodeElement).forEach((nodeData) => {
-                if (nodeData.nodeType === 1) {
-                  node = this.processVariable(
-                    nodeData, node, xmlDoc, protocol.codebook.node[nodeType], nodeType);
-                }
-              });
-              session.data.nodes.push(node);
-              break;
-            // eslint-disable-next-line no-case-declarations
-            case 'edge':
-              const edgeElement = entityElements[i].childNodes;
-              let edge = {};
-              edge.attributes = {};
-              const edgeType = this.lookUpEntityType(edgeElement, protocol.codebook.edge);
-              Array.from(edgeElement).forEach((edgeData) => {
-                if (edgeData.nodeType === 1) {
-                  edge = this.processVariable(
-                    edgeData, edge, xmlDoc, protocol.codebook.edge[edgeType], edgeType);
-                }
-              });
-              session.data.edges.push(edge);
-              break;
-            default:
-              break;
-          }
-        }
-      }
-      sessions.push(session);
-    });
-
-    return { protocolId, sessions };
-  }
-
-  // this is a string (name) in graphml, but uuid in export from NC to Server
-  lookUpEntityType = (entityElement, codebookEntity) => {
-    let typeUUID = '';
-
-    Array.from(entityElement).forEach((entityData) => {
-      if (entityData.nodeType === 1) {
-        const keyValue = entityData.getAttributeNode('key').value;
-        if (keyValue === 'networkCanvasType') {
-          typeUUID = Object.keys(codebookEntity).find(
-            key => codebookEntity[key].name === entityData.textContent);
-          typeUUID = typeUUID || entityData.textContent;
-        }
-      }
-    });
-    return typeUUID;
-  };
-
-  processVariable = (element, entity, xmlDoc, codebookEntity, entityType = '') => {
-    let keyValue = element.getAttributeNode('key').value;
-    if (keyValue === 'networkCanvasUUID') {
-      // eslint-disable-next-line no-underscore-dangle
-      return { ...entity, _uid: element.textContent };
-    } else if (keyValue === 'label') {
-      // can ignore since this was just for gephi
-      return entity;
-    } else if (keyValue === 'networkCanvasType') {
-      return { ...entity, type: entityType };
-    } else if (keyValue === 'networkCanvasSourceUUID') {
-      return { ...entity, from: element.textContent };
-    } else if (keyValue === 'networkCanvasTargetUUID') {
-      return { ...entity, to: element.textContent };
-    } else if (!keyValue.includes('_')) {
-      const graphmlKey = xmlDoc.getElementById(keyValue);
-      let text = element.textContent;
-      switch (graphmlKey.getAttributeNode('attr.type').value) {
-        case 'int':
-        case 'double':
-        case 'float':
-          text = !Number.isNaN(Number(text)) ? Number(text) : text;
-          break;
-        case 'boolean':
-          text = (text === 'true');
-          break;
-        case 'string':
-        default:
-          break;
-      }
-      // variables not in the codebook are external variables - use the name instead of uuid
-      keyValue = codebookEntity.variables[keyValue] ? keyValue : graphmlKey.getAttributeNode('attr.name').value;
-      return { ...entity, attributes: { ...entity.attributes, [keyValue]: text } };
-    } else if (keyValue.endsWith('_X')) { // process locations
-      const locationKey = keyValue.substring(0, keyValue.indexOf('_X'));
-      const locationObject = (entity.attributes && entity.attributes[locationKey]) || {};
-      const text = Number(element.textContent);
-      return {
-        ...entity,
-        attributes: { ...entity.attributes, [locationKey]: { ...locationObject, x: text } } };
-    } else if (keyValue.endsWith('_Y')) { // process locations
-      const locationKey = keyValue.substring(0, keyValue.indexOf('_Y'));
-      const locationObject = (entity.attributes && entity.attributes[locationKey]) || {};
-      const text = Number(element.textContent);
-      return {
-        ...entity,
-        attributes: { ...entity.attributes, [locationKey]: { ...locationObject, y: text } } };
-    }
-
-    // process categorical vars
-    if (element.textContent === 'true') {
-      const catKey = keyValue.substring(0, keyValue.indexOf('_'));
-      const catVar = (entity.attributes && entity.attributes[catKey]) || []; // previous options
-      const codebookVarName = codebookEntity.variables[catKey].name;
-      const catValue = xmlDoc.getElementById(keyValue).getAttributeNode('attr.name').value; // variable_option
-      const optionIndex = codebookVarName.length + 1; // add one for the underscore
-      // fallback to using whatever it after the first underscore
-      const codebookOptionName = optionIndex > 0 ? catValue.substring(optionIndex) : catValue.substring(catValue.indexOf('_') + 1);
-      catVar.push(codebookOptionName);
-      return { ...entity, attributes: { ...entity.attributes, [catKey]: catVar } };
-    }
-    return entity;
   }
 
   /**
